change project package names & alter REST API name
