--- WEB-MAC-HACK-KI/src/main/java/amigo/com/SwaggerConfig.java
@@ -1,26 +1,0 @@
-package amigo.com;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-/**
- * Created by NEXT on 2017. 11. 11..
- */
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("amigo.com.web"))
-                .paths(PathSelectors.any())
-                .build();
-
-    }
-}
