--- WEB-MAC-HACK-KI/src/main/java/amigo/com/domain/ChattingRoom.java
@@ -1,18 +1,0 @@
-package amigo.com.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-import javax.persistence.*;
-import java.util.Set;
-
-/**
- * Created by Naver on 2017. 11. 9..
- */
-@Entity
-public class ChattingRoom {
-    @Id
-    private long id;
-
-    @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
-    private Set<User> users;
-}
