--- WEB-MAC-HACK-KI/src/main/java/amigo/com/domain/PartyGuestRepository.java
@@ -1,9 +1,0 @@
-package amigo.com.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-/**
- * Created by Naver on 2017. 10. 18..
- */
-public interface PartyGuestRepository extends CrudRepository<PartyGuest, Long> {
-}
