--- WEB-MAC-HACK-KI/src/main/java/amigo/com/domain/UserRepository.java
@@ -1,10 +1,0 @@
-package amigo.com.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-/**
- * Created by Naver on 2017. 11. 8..
- */
-public interface UserRepository extends CrudRepository<User, Long>{
-    User findByEmail(String email);
-}
