--- WEB-MAC-HACK-KI/src/main/java/amigo/com/web/Result.java
@@ -1,18 +1,0 @@
-package amigo.com.web;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- * Created by Naver on 2017. 11. 9..
- */
-
-@NoArgsConstructor
-@AllArgsConstructor
-public class Result {
-    @Getter
-    @Setter
-    private String url;
-}
