+++ WEB-MAC-HACK-KI/src/main/java/com/amigotrip/domain/PartyGuestRepository.java
@@ -1,0 +1,9 @@
+package com.amigotrip.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+/**
+ * Created by Naver on 2017. 10. 18..
+ */
+public interface PartyGuestRepository extends CrudRepository<PartyGuest, Long> {
+}
