+++ WEB-MAC-HACK-KI/src/main/java/com/amigotrip/domain/RoleRepository.java
@@ -1,0 +1,10 @@
+package com.amigotrip.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+/**
+ * Created by Naver on 2017. 11. 15..
+ */
+public interface RoleRepository extends CrudRepository<Role, Long> {
+    Role findByRole(String role);
+}
