+++ WEB-MAC-HACK-KI/src/main/java/com/amigotrip/domain/UserRepository.java
@@ -1,0 +1,10 @@
+package com.amigotrip.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+/**
+ * Created by Naver on 2017. 11. 8..
+ */
+public interface UserRepository extends CrudRepository<User, Long>{
+    User findByEmail(String email);
+}
