+++ WEB-MAC-HACK-KI/src/main/java/com/amigotrip/service/CustomUserDetailsService.java
@@ -1,0 +1,50 @@
+package com.amigotrip.service;
+
+import com.amigotrip.domain.Role;
+import com.amigotrip.domain.UserRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import javax.transaction.Transactional;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+
+/**
+ * Created by Naver on 2017. 11. 12..
+ */
+@Slf4j
+@Service("customUserDetailsService")
+public class CustomUserDetailsService implements UserDetailsService{
+    @Autowired
+    private UserRepository userRepository;
+
+    @Override
+    @Transactional
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException{
+        log.debug("load username: {}", email);
+        com.amigotrip.domain.User user = userRepository.findByEmail(email);
+        log.debug("loaded user: {}", user);
+        List<GrantedAuthority> authorities = buildUserAuthority(user.getRoles());
+        return buildUserForAuthentication(user, authorities);
+    }
+
+    private User buildUserForAuthentication(com.amigotrip.domain.User user, List<GrantedAuthority> authorities) {
+        return new User(user.getEmail(), user.getPassword(), true, true, true, true, authorities);
+    }
+
+    private List<GrantedAuthority> buildUserAuthority(Set<Role> roles) {
+        List<GrantedAuthority> authorities = new ArrayList<GrantedAuthority>(0);
+        for(Role role : roles) {
+            authorities.add(new SimpleGrantedAuthority(role.getRole()));
+        }
+        return authorities;
+    }
+}
