+++ WEB-MAC-HACK-KI/src/main/java/com/amigotrip/service/UserConfirmService.java
@@ -1,0 +1,42 @@
+package com.amigotrip.service;
+
+import com.amigotrip.domain.RoleRepository;
+import com.amigotrip.domain.User;
+import com.amigotrip.domain.UserRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import javax.annotation.Resource;
+
+/**
+ * Created by Naver on 2017. 11. 15..
+ */
+@Slf4j
+@Service("service.userConfirmService")
+public class UserConfirmService {
+    @Resource
+    private UserRepository userRepository;
+
+    @Resource
+    private RoleRepository roleRepository;
+
+    public boolean isConfirmedUserByUserId(long userId) {
+        User user = userRepository.findOne(userId);
+        return user.getRoles().contains(roleRepository.findByRole("user"));
+    }
+
+    public String confirmUserByUserId(long userId, String key) {
+        User user = userRepository.findOne(userId);
+        if(user == null) {
+            return "redirect:/";
+        }
+
+        if(!isConfirmedUserByUserId(userId) && user.getEmailConfirmKey().equals(key)) {
+            user.addRole(roleRepository.findByRole("user"));
+            user.removeRole(roleRepository.findByRole("unconfirmed_user"));
+            userRepository.save(user);
+            return "redirect:/loginForm";
+        }
+        return "redirect:/";
+    }
+}
