+++ WEB-MAC-HACK-KI/src/test/java/com/amigotrip/BcryptTest.java
