+++ WEB-MAC-HACK-KI/src/test/java/com/amigotrip/mail/AmigoMailSenderTest.java
